pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "A minimal FastAPI project scaffold."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "langgraph>=0.6.7",
    "ag-ui-langgraph>=0.0.25",
    "langchain-community>=0.2.10",
    "python-dotenv>=1.0.1",
    "dashscope>=1.21.0",
    "deepagents>=0.4.8",
    "langgraph-checkpoint>=4.0.1",
    "langgraph-checkpoint-redis>=0.4.0",
    "langchain-deepseek>=1.0.1",
    "mem0ai>=1.0.7",
    "pymilvus>=2.6.10",
    "pyjwt>=2.12.1",
    "langchain-mcp-adapters>=0.2.2",
    "mcp>=1.26.0",
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    "langfuse>=4.0.6",
    "openpyxl>=3.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
